Made it possible to train model without connecting microbit
