+++ src/menus/ModelMenu.svelte
@@ -6,7 +6,7 @@
 
 <script lang="ts">
   import { state } from '../script/stores/uiStore';
-  import { bestPrediction } from '../script/stores/mlStore';
+  import { GestureData, bestPrediction } from '../script/stores/mlStore';
   import { t } from '../i18n';
 
   $: confidence = $state.isInputReady
@@ -14,8 +14,18 @@
     : 0;
   confidence = isNaN(confidence) ? 0 : confidence;
 
+  const getPredictionLabel = (isInputReady: boolean, bestPrediction?: GestureData) => {
+    if (!isInputReady) {
+      return $t('menu.model.connectInputMicrobit');
+    }
+    if (bestPrediction) {
+      return bestPrediction.name
+    }
+    return '...';
+  }
+
   $: confidenceLabel = Math.round(confidence * 100).toString() + '%';
-  $: predictionLabel = !$state.isInputReady ? '' : $bestPrediction?.name ?? '';
+  $: predictionLabel = getPredictionLabel($state.isInputReady, $bestPrediction);
 </script>
 
 <div class="w-full text-center justify-center pt-5">
@@ -31,7 +41,9 @@
   {:else}
     <div
       class="grid break-words mr-auto ml-auto w-3/4 h-70px border-2 rounded-lg border-solid text-center align-center content-center">
-      <p class="w-full max-w-[100%] text-2xl break-all">
+      <p class="w-full max-w-[100%] text-2xl break-all"
+      class:text-2xl={$state.isInputReady}
+      class:text-md={!$state.isInputReady}>
         {predictionLabel}
       </p>
     </div>
