+++ src/pages/model/ModelPage.svelte
@@ -10,8 +10,8 @@
   import ControlBar from '../../components/control-bar/ControlBar.svelte';
   import ExpandableControlBarMenu from '../../components/control-bar/control-bar-items/ExpandableControlBarMenu.svelte';
   import { ModelView, state } from '../../script/stores/uiStore';
-  import ModelPageStackView from './ModelPageStackView.svelte';
-  import ModelPageTileView from './ModelPageTileView.svelte';
+  import ModelPageStackView from './stackview/ModelPageStackView.svelte';
+  import ModelPageTileView from './tileview/ModelPageTileView.svelte';
 
   const openMakeCodeInNewTab = () => {
     window.open(
@@ -30,6 +30,7 @@
     </ExpandableControlBarMenu>  
   </ControlBar>
 </div>
+
 <div class="pt-4 pl-3">
   {#if $state.modelView == ModelView.TILE}
     <ModelPageTileView />
